Refactor: Document 폴더 DocumentList 폴더 내부로 이동
